ENH: Improve PR workflows (#697)
* Run `pytest` and `pytest` on GPUs in parallel rather than sequentially.
* Cancel existing runs from previous pushes
* Fix dependency problem in hi-ml PR
* Reduce logging noise from `numba`
