+++ .github/actions/cancel_azureml_jobs/cancel_aml_jobs.py
@@ -36,7 +36,12 @@ def cancel_running_and_queued_jobs() -> None:
             print(f"Skipping {status_suffix}")
         else:
             print(f"Cancelling {status_suffix}")
-            run.cancel()
+            try:
+                run.cancel()
+            except Exception as ex:
+                # Exceptions here are rare, but do happen. Sometimes AML says "Run is in a terminal state", and
+                # refuses to cancel.
+                print(f"Unable to cancel {status_suffix}: {ex}")
 
 
 if __name__ == "__main__":
