+++ .github/workflows/cpath-pr.yml
@@ -13,10 +13,14 @@ on:
       - "hi-ml-azure/**"
       - "hi-ml/**"
 
+# Cancel previous runs of this workflow that are still in progress.
+concurrency:
+  group: ${{ github.ref }}/cpath-pr
+  cancel-in-progress: ${{ github.ref != 'refs/heads/master' }}
+
 env:
   pythonVersion: 3.9
   folder: hi-ml-cpath
-  module_for_coverage_reporting: health_cpath
   HIML_TENANT_ID: ${{ secrets.HIML_TENANT_ID }}
   HIML_RESOURCE_GROUP: ${{ secrets.HIML_RESOURCE_GROUP }}
   HIML_SUBSCRIPTION_ID: ${{ secrets.HIML_SUBSCRIPTION_ID }}
@@ -75,7 +79,6 @@ jobs:
 
   pytest:
     runs-on: ubuntu-20.04
-    needs: [ cancel-azureml ]
     steps:
       - uses: actions/checkout@v3
         with:
@@ -92,16 +95,33 @@ jobs:
           make pip_test
           make pytest_coverage
 
-      - name: Run GPU tests
-        # GPU tests should be run even if other tests fail
+      - name: Upload coverage reports to Codecov
+        # Coverage should also be uploaded if tests still fail.
+        # This will pick up the file coverage.xml from the test run
         if: always()
-        shell: bash -l {0}
+        uses: codecov/codecov-action@v3
+        with:
+            flags: ${{ env.folder }}
+
+  pytest_gpu:
+    runs-on: ubuntu-20.04
+    needs: [ cancel-azureml ]
+    steps:
+      - uses: actions/checkout@v3
+        with:
+          lfs: true
+
+      - name: Prepare Conda environment
+        uses: ./.github/actions/prepare_cpath_environment
+
+      - name: Run GPU tests
         run: |
-          python hi-ml-azure/run_pytest.py --mark=gpu --cluster=pr-gpu --conda_env=${{ env.folder }}/environment.yml --folder=${{ env.folder }} --coverage_module=${{ env.module_for_coverage_reporting }}
+          cd ${{ env.folder }}
+          make pytest_gpu
 
       - name: Upload coverage reports to Codecov
         # Coverage should also be uploaded if tests still fail.
-        # This will pick up both the files coverage.xml (from the normal tests) and pytest_gpu_coverage.xml (from the GPU tests).
+        # This will pick up the file pytest_gpu_coverage.xml from the GPU tests.
         if: always()
         uses: codecov/codecov-action@v3
         with:
@@ -289,6 +309,7 @@ jobs:
       flake8,
       mypy,
       pytest,
+      pytest_gpu,
       smoke_test_cucim_slidespandaimagenetmil,
       smoke_test_openslide_slidespandaimagenetmil,
       smoke_test_tilespandaimagenetmil,
