+++ .github/workflows/hi-ml-pr.yml
@@ -11,6 +11,11 @@ on:
       - "hi-ml/**"
       - ".github/workflows/hi-ml-pr.yml"
 
+# Cancel previous runs of this workflow that are still in progress.
+concurrency:
+  group: ${{ github.ref }}/hi-ml-pr
+  cancel-in-progress: ${{ github.ref != 'refs/heads/master' }}
+
 env:
   pythonVersion: 3.7
   HIML_TENANT_ID: ${{ secrets.HIML_TENANT_ID }}
@@ -216,7 +221,7 @@ jobs:
 
   publish-testpypi-pkg:
     runs-on: ubuntu-20.04
-    needs: [ test-artifact-pkg ]
+    needs: [ test-artifact-pkg, pyright, mypy, flake8 ]
     if: "!startsWith(github.ref, 'refs/tags/v')"
     strategy:
       matrix:
@@ -248,7 +253,7 @@ jobs:
 
   publish-pypi-pkg:
     runs-on: ubuntu-20.04
-    needs: [ test-artifact-pkg ]
+    needs: [ test-artifact-pkg, pyright, mypy, flake8 ]
     strategy:
       matrix:
         folder: [ hi-ml, hi-ml-azure ]
